Add files via upload
This project made during the Summer semester. We have been in only 4 meets and then the war started... I was recruited to the army soon as the war started and that is the reason that the Fire Truck robot does not respond as well by the sensors in the movie. After the first match I started over again by myself to rewrite the code in better way which I was thinking about when I was in rest time inside Gaza with no any electronic device.
This is the first version on the code that shown in the video, But I was uploaded the better one here on Github. enjoy! 

https://youtu.be/-p2JRCPtoWk?feature=shared