Create README.md
The code shown here is from a course during my mechanical engineering degree in designing a fire engine robot with a Lego kit, sensors and motors. Beyond the mechanical engineering field, there is also, of course, the field of developing a code and control algorithm for the sensors due to the need for an automatic robot to put out fires. Basically, the robot operates automatically, maneuvers between obstacles, and when a fire is detected, it knows how to position itself and direct the water injector in the direction of the fire.

The code shown here is from a course during my mechanical engineering degree in designing a fire engine robot with a Lego kit, sensors and motors. Beyond the mechanical engineering field, there is also, of course, the field of developing a code and control algorithm for the sensors due to the need for an automatic robot to put out fires. Basically, the robot operates automatically, maneuvers between obstacles, and when a fire is detected, it knows how to position itself and direct the water injector in the direction of the fire.